Put OnePass class back inside lookup plugin
There is no good place for sharing code across lookups currently.
